+++ tests/integ/test_dataframe.py
@@ -15,7 +15,6 @@ from itertools import product
 from typing import Tuple
 from unittest import mock
 
-from snowflake.snowpark.dataframe import map
 from snowflake.snowpark.session import Session
 from tests.conftest import local_testing_mode
 
@@ -1677,6 +1676,19 @@ def test_create_dataframe_with_dict(session):
             Row(None, None, None, 4, 5),
         ],
     )
+
+
+def test_dataframe_transform_negative(session):
+    def ret_int_fn(input_df):
+        return 2
+
+    df = session.create_dataframe([[1, 2]], schema=["a", "b"])
+
+    with pytest.raises(
+        AssertionError,
+        match="Expected return value to be an instance of class DataFrame, got <class 'int'> instead.",
+    ):
+        df.transform(ret_int_fn)
 
 
 def test_create_dataframe_with_dict_given_schema(session):
@@ -4341,12 +4353,70 @@ def test_map_basic(
         row = Row(*output_col_names)
         expected = [row(*e) for e in expected]
         Utils.check_answer(
-            map(df, func, output_types, output_column_names=output_col_names), expected
+            df.map(func, output_types, output_column_names=output_col_names), expected
         )
     else:
         row = Row(*[f"c_{i+1}" for i in range(len(output_types))])
         expected = [row(*e) for e in expected]
-        Utils.check_answer(map(df, func, output_types), expected)
+        Utils.check_answer(df.map(func, output_types), expected)
+
+
+@pytest.mark.skipif(
+    "config.getoption('local_testing_mode', default=False)",
+    reason="Table function is not supported in Local Testing",
+)
+@pytest.mark.udf
+@pytest.mark.parametrize("overlapping_columns", [True, False])
+@pytest.mark.parametrize(
+    "func,output_types,output_column_names,expected",
+    [
+        (
+            lambda row: (row[1] + 1, row[1] + 2),
+            [IntegerType()],
+            ["A"],
+            [(i + 1,) for i in range(5)] + [(i + 2,) for i in range(5)],
+        ),
+        (
+            lambda row: [(row.B * 2, row.C)],
+            [IntegerType(), StringType()],
+            ["B", "C"],
+            [
+                (
+                    i * 2,
+                    f"w{i}",
+                )
+                for i in range(5)
+            ],
+        ),
+        (
+            lambda row: [
+                Row(row.B * row.B, f"-{row.C}"),
+                Row(row.B + row.B, f"+{row.C}"),
+            ],
+            [IntegerType(), StringType()],
+            ["B", "C"],
+            [(i * i, f"-w{i}") for i in range(5)]
+            + [(i + i, f"+w{i}") for i in range(5)],
+        ),
+    ],
+)
+def test_flat_map_basic(
+    session, func, output_types, output_column_names, expected, overlapping_columns
+):
+    df = session.create_dataframe(
+        [(True, i, f"w{i}") for i in range(5)], schema=["a", "b", "c"]
+    )
+    if overlapping_columns:
+        row = Row(*output_column_names)
+        expected = [row(*e) for e in expected]
+        Utils.check_answer(
+            df.flat_map(func, output_types, output_column_names=output_column_names),
+            expected,
+        )
+    else:
+        row = Row(*[f"C_{i+1}" for i in range(len(output_types))])
+        expected = [row(*e) for e in expected]
+        Utils.check_answer(df.flat_map(func, output_types), expected)
 
 
 @pytest.mark.skipif(
@@ -4386,7 +4456,75 @@ def test_map_vectorized(session, func, output_types, expected):
     )
 
     Utils.check_answer(
-        map(df, func, output_types, vectorized=True, packages=["pandas"]), expected
+        df.map(func, output_types, vectorized=True, packages=["pandas"]), expected
+    )
+
+
+@pytest.mark.skipif(
+    "config.getoption('local_testing_mode', default=False)",
+    reason="Table function is not supported in Local Testing",
+)
+@pytest.mark.skipif(not is_pandas_available, reason="pandas is required for this test")
+@pytest.mark.udf
+@pytest.mark.parametrize(
+    "func,data,output_types,partition_by,expected_lambda",
+    [
+        (lambda df: df, [["a"], ["b"], ["c"]], [StringType()], None, lambda row: [row]),
+        (
+            lambda df: df.explode("_1"),
+            [
+                [["A", "B", "C"], "Guard", 7],
+                [["D"], "Forward", 15],
+                [["E", "F"], "Center", 20],
+            ],
+            [StringType(), StringType(), IntegerType()],
+            None,
+            lambda row: [(row[0][i], row[1], row[2]) for i in range(len(row[0]))],
+        ),
+        (
+            lambda df: df.explode("_1"),
+            [
+                [["A", "B", "C"], "Guard", 7],
+                [["D"], "Forward", 15],
+                [["E", "F"], "Center", 20],
+            ],
+            [StringType(), StringType(), IntegerType()],
+            "_2",
+            lambda row: [(row[0][i], row[1], row[2]) for i in range(len(row[0]))],
+        ),
+        (
+            lambda df: ((len(df.iloc[0]["_1"]),), (df.iloc[0]["_2"],)),
+            [
+                [["A", "B", "C"], "Guard", 7],
+                [["D"], "Forward", 15],
+                [["E", "F"], "Center", 20],
+            ],
+            [
+                IntegerType(),
+                StringType(),
+            ],
+            "_2",
+            lambda row: [(len(row[0]), row[1])],
+        ),
+    ],
+)
+def test_flat_map_vectorized(
+    session, func, data, output_types, partition_by, expected_lambda
+):
+    df = session.create_dataframe(data=data)
+    expected = []
+    for row in data:
+        expected.extend(expected_lambda(row))
+    print(expected)
+    Utils.check_answer(
+        df.flat_map(
+            func,
+            output_types,
+            vectorized=True,
+            partition_by=partition_by,
+            packages=["pandas"],
+        ),
+        expected,
     )
 
 
@@ -4400,12 +4538,10 @@ def test_map_chained(session):
         [[True, i, f"w{i}"] for i in range(5)], schema=["A", "B", "C"]
     )
 
-    new_df = map(
-        map(
-            df,
-            lambda x: (x.B * x.B, f"_{x.C}_"),
-            output_types=[IntegerType(), StringType()],
-        ),
+    new_df = df.map(
+        lambda x: (x.B * x.B, f"_{x.C}_"),
+        output_types=[IntegerType(), StringType()],
+    ).map(
         lambda x: len(x[1]) + x[0],
         output_types=[IntegerType()],
     )
@@ -4414,13 +4550,11 @@ def test_map_chained(session):
     Utils.check_answer(new_df, expected)
 
     # chained calls with repeated column names
-    new_df = map(
-        map(
-            df,
-            lambda x: Row(x.B * x.B, f"_{x.C}_"),
-            output_types=[IntegerType(), StringType()],
-            output_column_names=["A", "B"],
-        ),
+    new_df = df.map(
+        lambda x: Row(x.B * x.B, f"_{x.C}_"),
+        output_types=[IntegerType(), StringType()],
+        output_column_names=["A", "B"],
+    ).map(
         lambda x: Row(len(x.B) + x.A),
         output_types=[IntegerType()],
         output_column_names=["A"],
@@ -4429,21 +4563,65 @@ def test_map_chained(session):
     Utils.check_answer(new_df, expected)
 
 
-def test_map_negative(session):
+@pytest.mark.skipif(
+    "config.getoption('local_testing_mode', default=False)",
+    reason="Table function is not supported in Local Testing",
+)
+@pytest.mark.skipif(not is_pandas_available, reason="pandas is required for this test")
+@pytest.mark.udf
+def test_flat_map_chained(session):
+    # explode, partitiiton by _2 and apply len
+    data = [
+        [["A", "B", "C"], "Guard", 7],
+        [["D"], "Forward", 15],
+        [["E", "F"], "Center", 20],
+    ]
+    schema = ["Team", "Position", "Points"]
+    df = session.create_dataframe(data, schema=schema)
+    new_df = df.flat_map(
+        lambda pdf: pdf.explode("TEAM"),
+        output_types=[StringType(), StringType(), IntegerType()],
+        output_column_names=schema,
+        vectorized=True,
+        packages=["pandas"],
+    ).flat_map(
+        lambda pdf: (
+            (len(pdf),),
+            (pdf.iloc[0]["POSITION"],),
+        ),
+        output_types=[IntegerType(), StringType()],
+        output_column_names=["SIZE", "POSITION"],
+        vectorized=True,
+        partition_by="POSITION",
+    )
+    Utils.check_answer(new_df, [(3, "Guard"), (1, "Forward"), (2, "Center")])
+
+
+@pytest.mark.parametrize("test_flat_map", [True, False])
+def test_map_negative(session, test_flat_map):
     df1 = session.create_dataframe(
         [[True, i, f"w{i}"] for i in range(5)], schema=["A", "B", "C"]
     )
 
     with pytest.raises(ValueError, match="output_types cannot be empty."):
-        map(df1, lambda row: [row.B, row.C], output_types=[])
+        if test_flat_map:
+            df1.flat_map(lambda row: [row.B, row.C], output_types=[])
+        else:
+            df1.map(lambda row: [row.B, row.C], output_types=[])
 
     with pytest.raises(
         ValueError,
         match="'output_column_names' and 'output_types' must be of the same size.",
     ):
-        map(
-            df1,
-            lambda row: [row.B, row.C],
-            output_types=[IntegerType(), StringType()],
-            output_column_names=["a", "b", "c"],
-        )
+        if test_flat_map:
+            df1.flat_map(
+                lambda row: [row.B, row.C],
+                output_types=[IntegerType(), StringType()],
+                output_column_names=["a", "b", "c"],
+            )
+        else:
+            df1.map(
+                lambda row: [row.B, row.C],
+                output_types=[IntegerType(), StringType()],
+                output_column_names=["a", "b", "c"],
+            )